Move 'geodesic' from product manifold to metric
